# openapi.yaml
openapi: 3.1.0
info:
  title: Cloud Commitment Portfolio Optimizer API
  version: 0.1.0
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    JwtBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AnalystApiKey:
      type: apiKey
      in: header
      name: X-API-Key
    AccessCookie:
      type: apiKey
      in: cookie
      name: __Host-ccpo_access
      description: Production HttpOnly RS256 access cookie. Loopback development/test uses ccpo_access.
    RefreshCookie:
      type: apiKey
      in: cookie
      name: __Host-ccpo_refresh
      description: Production HttpOnly opaque refresh cookie. Loopback development/test uses ccpo_refresh.
    CsrfCookie:
      type: apiKey
      in: cookie
      name: __Host-ccpo_csrf
      description: Production readable session-bound CSRF cookie. Loopback development/test uses ccpo_csrf.
  headers:
    RequestId:
      description: Server-generated request correlation identifier.
      schema: { type: string }
    NoStore:
      description: Prevents storage of sensitive or tenant-scoped responses.
      schema: { type: string, const: no-store }
    RetryAfter:
      description: Whole seconds before another attempt.
      schema: { type: string, pattern: "^[1-9][0-9]*$" }
    SetCookie:
      description: Exact Secure/Strict/Path=/ host cookie in production or explicit non-Secure loopback variant. Values are never documented or logged.
      schema: { type: string, writeOnly: true }
  schemas:
    TenantRegistrationAddress:
      type: object
      additionalProperties: false
      properties:
        line1: { type: string, maxLength: 200 }
        line2: { type: string, maxLength: 200 }
        locality: { type: string, maxLength: 100 }
        region: { type: string, maxLength: 100 }
        postal_code: { type: string, maxLength: 32 }
        country_code: { type: string, pattern: "^[A-Za-z]{2}$" }
    TenantRegistrationRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, maxLength: 200 }
        legal_name: { type: string, maxLength: 200 }
        full_legal_name: { type: string, maxLength: 200 }
        display_name: { type: string, maxLength: 200 }
        address: { $ref: "#/components/schemas/TenantRegistrationAddress" }
        registration:
          type: object
          maxProperties: 32
          additionalProperties: { type: string, maxLength: 256 }
        contact_email: { type: string, maxLength: 254 }
        contact_phone: { type: string, maxLength: 64 }
        support_url: { type: string, format: uri, maxLength: 2048 }
        finance_owner_email: { type: string, maxLength: 254 }
        wordmark: { type: string, maxLength: 256 }
        default_currency: { type: string, pattern: "^[A-Za-z]{3}$" }
        timezone: { type: string, maxLength: 64 }
        risk_budget_cents:
          type: string
          pattern: "^(0|[1-9][0-9]*)$"
          x-maximum: "9223372036854775807"
    TenantProfile:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - legal_name
        - full_legal_name
        - display_name
        - address
        - registration
        - contact_email
        - contact_phone
        - support_url
        - finance_owner_email
        - wordmark
        - default_currency
        - timezone
        - risk_budget_cents
        - is_active
        - created_at
        - updated_at
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        legal_name: { type: string }
        full_legal_name: { type: string }
        display_name: { type: string }
        address: { $ref: "#/components/schemas/TenantRegistrationAddress" }
        registration: { type: object, additionalProperties: { type: string } }
        contact_email: { type: [string, "null"] }
        contact_phone: { type: [string, "null"] }
        support_url: { type: [string, "null"] }
        finance_owner_email: { type: [string, "null"] }
        wordmark: { type: [string, "null"] }
        default_currency: { type: string, pattern: "^[A-Z]{3}$" }
        timezone: { type: string }
        risk_budget_cents:
          type: string
          pattern: "^(0|[1-9][0-9]*)$"
          x-maximum: "9223372036854775807"
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    TenantRegistrationCreated:
      type: object
      additionalProperties: false
      required: [tenant, apiKey]
      properties:
        tenant: { $ref: "#/components/schemas/TenantProfile" }
        apiKey:
          type: string
          description: Plaintext analyst automation credential returned once after commit.
    RegistrationError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, details]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                additionalProperties: false
                properties:
                  tenant_id: { type: string, format: uuid }
                  api_key_id: { type: string, format: uuid }
...
    UserRole:
      type: string
      enum: [tenant_admin, finops_analyst, finance_approver, read_only_auditor]
...
    AuthLoginRequest:
      type: object
      additionalProperties: false
      required: [tenant_id, email, password]
      properties:
        tenant_id: { type: string, format: uuid }
        email: { type: string, maxLength: 254 }
        password:
          type: string
          writeOnly: true
          description: NFC-normalized without trim; 15..128 Unicode scalars and at most 512 UTF-8 bytes.
    AuthSessionMetadata:
      type: object
      additionalProperties: false
      required: [user_id, tenant_id, role, access_expires_at, refresh_idle_expires_at, refresh_absolute_expires_at]
      properties:
        user_id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        role: { $ref: "#/components/schemas/UserRole" }
        access_expires_at: { type: string, format: date-time }
        refresh_idle_expires_at: { type: string, format: date-time }
        refresh_absolute_expires_at: { type: string, format: date-time }
    AuthSessionResponse:
      type: object
      additionalProperties: false
      required: [session]
      properties:
        session: { $ref: "#/components/schemas/AuthSessionMetadata" }
    AuthSessionError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, details]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: array, maxItems: 0 }
...
paths:
  /api/auth/login:
    post:
      operationId: createBrowserSession
      summary: Prove one tenant-local password and create a cookie-only session
      description: Credential headers are forbidden; stale cookies are ignored. Redis atomically admits account and immediate-peer-derived IP buckets before exactly one bounded Argon2id verification.
      security: []
      parameters:
        - name: Origin
          in: header
          required: true
          schema: { type: string, format: uri }
        - name: Sec-Fetch-Site
          in: header
          required: false
          schema: { type: string, const: same-origin }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthLoginRequest" }
      responses:
        "200":
          description: Exact non-secret session metadata after family, root token, and audit commit; three cookie values are transport-only.
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
            Cache-Control: { $ref: "#/components/headers/NoStore" }
            Set-Cookie: { $ref: "#/components/headers/SetCookie" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSessionResponse" }
        "400": { description: Closed JSON, media, password, or field validation failed., content: { application/json: { schema: { $ref: "#/components/schemas/AuthSessionError" } } } }
        "401": { description: Generic invalid proof or forbidden credential-header conflict., content: { application/json: { schema: { $ref: "#/components/schemas/AuthSessionError" } } } }
        "403": { description: Same-origin checks failed or valid proof found inactive user or tenant., content: { application/json: { schema: { $ref: "#/components/schemas/AuthSessionError" } } } }
        "413": { description: Encoded body exceeds 4096 bytes., content: { application/json: { schema: { $ref: "#/components/schemas/AuthSessionError" } } } }
        "429":
          description: Atomic account or canonical-IP rolling bucket denied admission.
          headers: { Retry-After: { $ref: "#/components/headers/RetryAfter" } }
          content: { application/json: { schema: { $ref: "#/components/schemas/AuthSessionError" } } }
        "503": { description: Redis, Argon worker, PostgreSQL, audit, or signing dependency unavailable., content: { application/json: { schema: { $ref: "#/components/schemas/AuthSessionError" } } } }
  /api/auth/refresh:
    post:
      operationId: rotateBrowserSession
      summary: Rotate the current opaque refresh token in its stable family
      description: No decoded body or credential headers. The IP bucket runs before lookup and the confirmed-family bucket before the family-lock transaction. Routine success writes no audit.
      security:
        - RefreshCookie: []
          CsrfCookie: []
      parameters:
        - name: Origin
          in: header
          required: true
          schema: { type: string, format: uri }
        - name: Sec-Fetch-Site
          in: header
          required: false
          schema: { type: string, const: same-origin }
        - name: X-CSRF-Token
          in: header
          required: true
          schema: { type: string, minLength: 43, maxLength: 43, writeOnly: true }
      responses:
        "200":
          description: Exact non-secret session metadata after current-token use and one child commit; three rotated cookie values are transport-only.
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
            Cache-Control: { $ref: "#/components/headers/NoStore" }
            Set-Cookie: { $ref: "#/components/headers/SetCookie" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSessionResponse" }
        "400": { description: A decoded non-empty body is forbidden., content: { application/json: { schema: { $ref: "#/components/schemas/AuthSessionError" } } } }
        "401": { description: Missing proof or generic malformed unknown expired used or revoked refresh., content: { application/json: { schema: { $ref: "#/components/schemas/AuthSessionError" } } } }
        "403": { description: Same-origin or CSRF failed or valid proof found inactive user or tenant., content: { application/json: { schema: { $ref: "#/components/schemas/AuthSessionError" } } } }
        "413": { description: Transport body exceeds 1024 bytes., content: { application/json: { schema: { $ref: "#/components/schemas/AuthSessionError" } } } }
        "429":
          description: Canonical-IP or confirmed-family rolling bucket denied admission.
          headers: { Retry-After: { $ref: "#/components/headers/RetryAfter" } }
          content: { application/json: { schema: { $ref: "#/components/schemas/AuthSessionError" } } }
        "503": { description: Redis PostgreSQL audit or signing dependency unavailable., content: { application/json: { schema: { $ref: "#/components/schemas/AuthSessionError" } } } }
  /api/auth/logout:
    post:
      operationId: revokeBrowserSession
      summary: Idempotently clear cookies and revoke a known stable family
      description: No decoded body or credential headers. Missing malformed unknown or already-revoked refresh state remains audit-free 204; a known unrevoked family commits one logout audit before 204.
      security:
        - {}
        - RefreshCookie: []
          CsrfCookie: []
      parameters:
        - name: Origin
          in: header
          required: false
          schema: { type: string, format: uri }
        - name: Sec-Fetch-Site
          in: header
          required: false
          schema: { type: string, const: same-origin }
        - name: X-CSRF-Token
          in: header
          required: false
          schema: { type: string, minLength: 43, maxLength: 43, writeOnly: true }
      responses:
        "204":
          description: Empty response after applicable transaction; all three exact cookies are cleared.
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
            Cache-Control: { $ref: "#/components/headers/NoStore" }
            Set-Cookie: { $ref: "#/components/headers/SetCookie" }
        "400": { description: A decoded non-empty body is forbidden., content: { application/json: { schema: { $ref: "#/components/schemas/AuthSessionError" } } } }
        "401": { description: Credential headers conflict with the endpoint contract., content: { application/json: { schema: { $ref: "#/components/schemas/AuthSessionError" } } } }
        "403": { description: Same-origin or CSRF proof failed for a presented refresh cookie., content: { application/json: { schema: { $ref: "#/components/schemas/AuthSessionError" } } } }
        "413": { description: Transport body exceeds 1024 bytes., content: { application/json: { schema: { $ref: "#/components/schemas/AuthSessionError" } } } }
        "503": { description: PostgreSQL or audit dependency unavailable; server revocation is not claimed., content: { application/json: { schema: { $ref: "#/components/schemas/AuthSessionError" } } } }
  /api/tenants/register:
    post:
      operationId: registerTenant
      summary: Register a tenant and issue its first analyst automation credential once
      description: Disabled by default. Successful credentials cannot be replayed or recovered by this public route.
      security: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 16
            maxLength: 128
            pattern: "^[!-~]+$"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TenantRegistrationRequest" }
      responses:
        "201":
          description: Tenant created; the plaintext apiKey is returned on this response only.
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
            Cache-Control: { $ref: "#/components/headers/NoStore" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TenantRegistrationCreated" }
        "400":
          description: Strict request, media, JSON, header, or normalization validation failed.
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
            Cache-Control: { $ref: "#/components/headers/NoStore" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RegistrationError" }
        "404":
          description: Self-registration is disabled and the resource behaves as absent.
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
            Cache-Control: { $ref: "#/components/headers/NoStore" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RegistrationError" }
        "409":
          description: Idempotency request is in progress, non-replayable, reused, or terminal.
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
            Cache-Control: { $ref: "#/components/headers/NoStore" }
            Retry-After:
              $ref: "#/components/headers/RetryAfter"
              description: Present only for IDEMPOTENCY_IN_PROGRESS and always 1.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RegistrationError" }
        "413":
          description: JSON request exceeds 16384 bytes.
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
            Cache-Control: { $ref: "#/components/headers/NoStore" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RegistrationError" }
        "429":
          description: Five attempts were already admitted for the canonical client IP in the rolling 60-second window.
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
            Cache-Control: { $ref: "#/components/headers/NoStore" }
            Retry-After: { $ref: "#/components/headers/RetryAfter" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RegistrationError" }
        "503":
          description: The required PostgreSQL or shared limiter dependency is unavailable.
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
            Cache-Control: { $ref: "#/components/headers/NoStore" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RegistrationError" }
...